requirements.txt
streamlit
opencv-python-headless
numpy
mediapipe
tensorflow
keras
python-dotenv
spotipy
